feat(code): 实现编码规则功能并优化相关代码
- 新增 vulcan-code 模块并集成到项目中
- 实现编码规则配置、生成、重置等功能
- 优化序列号生成逻辑，改为使用 Integer 类型
- 更新相关接口和方法以适应新的序列号生成逻辑
- 重构部分代码结构，提高可维护性

+++ vulcan-code/src/main/java/com/vulcan/code/controller/CodeRuleController.java
@@ -118,7 +118,7 @@ public class CodeRuleController {
      * @return 重置结果
      */
     @PutMapping("/{id}/sequence/{sequence}")
-    public ResultVo<Boolean> resetSequence(@PathVariable Long id, @PathVariable Long sequence) {
+    public ResultVo<Boolean> resetSequence(@PathVariable Long id, @PathVariable Integer sequence) {
         return ResultVo.success(codeRuleService.resetSequence(id, sequence));
     }
